fix: Resolve all TypeScript build errors
   - Fix NextAuth authOptions export issue by moving to separate config file
   - Fix AvatarImage src prop type handling in sidebar components
   - Fix globalThis.alertSettings null handling in websocketDb
   - Fix missing TEST_DEVICE_NAME constant in test script
   - Add proper TypeScript typing for NextAuth configuration

   Build now compiles successfully with no TypeScript errors.
