+++ Server/app/api/admin/alert-thresholds/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { getServerSession } from 'next-auth';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/lib/auth-config';
 import getDb from '@/lib/getDb';
 
 // GET - Fetch current alert thresholds
